Make sure XDG_RUNTIME_DIR is set in startplasma
Whilst X11 can get away without this, it's not ideal as the XDG spec
requires applications to warn if they have to fall back on a default,
and Wayland is built around it being set, since WAYLAND_DISPLAY is
relative to it. Whilst these days WAYLAND_DISPLAY can be an absolute
path, libwayland, and the libwayland-derived helpers used by
startplasma, don't let you do anything but create a relative socket, and
older clients won't understand it as it wasn't in the early versions.
Thus, define the variable here if it's not set rather than treading
further off the beaten path.

Note that, as far as I can tell, a full SDDM + ConsoleKit setup won't
get the variable defined, only an SDDM + (e)logind one, so I don't think
this is a problem specific to our cut-down environment but instead one
that also affects FreeBSD. Running with ck-launch-session, as you'd
normally do on the command line in order to be able to open devices
associated with your current seat, would set it though.

This change was imported from CheriBSD ports [1].

[1] https://github.com/CTSRD-CHERI/cheribsd-ports/commit/9ee16a242e1aadc5a4e4262347e77042598a34a6